WIP: moved jsns out of internal directory
Closes #5.
